Move all contract tests to folder
