+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Contracts/CosmosVirtualUnitTest.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation.  All rights reserved.
 //------------------------------------------------------------
 
-namespace Microsoft.Azure.Cosmos.Tests
+namespace Microsoft.Azure.Cosmos.Tests.Contracts
 {
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
